Exercise 4.21. -

+++ part4/blog/controllers/blogs.js
@@ -57,8 +57,27 @@ blogsRouter.put("/:id", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  const blog = await Blog.findById(request.params.id);
+
+  console.log("user: ", user.id.toString());
+
+  console.log("blog: ", blog);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    return response.status(204).end();
+  } else {
+    return response
+      .status(401)
+      .json({ error: "Unauthorized user. You may only delete your own blogs" });
+  }
 });
 
 module.exports = blogsRouter;
